ref: move types + parsers to helper files
